feat: implement NewsLetter component with form and styling; refactor page layout
