--- src/sections/NewsLetter/NewsLetter.js
@@ -1,15 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import styles from './NewsLetter.module.scss';
-
-const NewsLetter = () => (
-  <div className={styles.NewsLetter}>
-    NewsLetter Component
-  </div>
-);
-
-NewsLetter.propTypes = {};
-
-NewsLetter.defaultProps = {};
-
-export default NewsLetter;
